--- UI/src/components/templates/RipaHomeTemplate.vue
@@ -1,15 +1,0 @@
-<template>
-  <ripa-intro-wrapper></ripa-intro-wrapper>
-</template>
-
-<script>
-import RipaIntroWrapper from '@/components/organisms/RipaIntroWrapper'
-
-export default {
-  name: 'ripa-home-template',
-
-  components: {
-    RipaIntroWrapper,
-  },
-}
-</script>
