+++ UI/src/stories/containers/RipaHomeContainer.stories.js
@@ -11,7 +11,7 @@ export default {
       {},
       {
         routes: routesData,
-        initialEntry: '/',
+        initialEntry: '/form',
       },
     ),
   ],
